+++ crates/api/CHANGELOG.md
@@ -22,4 +22,4 @@
 
 ## 0.1.0
 
-<!-- Update PR number to skip CHANGELOG.md test: #0 -->
+<!-- Update PR number to skip CHANGELOG.md test: #112 -->